Add API for getting Helm Release
