+++ custom-src/frontend/app/custom/kubernetes/workloads/release/tabs/helm-release-helper.service.ts
@@ -13,7 +13,7 @@ import {
   GetHelmReleaseGraph,
   GetHelmReleaseResource,
   GetHelmReleases,
-  GetHelmReleaseStatus,
+  GetHelmRelease,
 } from '../../store/workloads.actions';
 import {
   HelmRelease,
@@ -46,23 +46,30 @@ export class HelmReleaseHelperService {
     this.namespace = this.guid.split(':')[1];
     this.endpointGuid = this.guid.split(':')[0];
 
-    // TODO: NWM This should use a `GetHelmRelease` action instead of fetching the entire list
-    const action = new GetHelmReleases();
-    const paginationMonitor = new PaginationMonitor(store, action.paginationKey, kubernetesEntityFactory(helmReleaseEntityKey));
-    const svc = getPaginationObservables({ store, action, paginationMonitor });
-    this.isFetching$ = svc.fetchingEntities$;
+    const action = new GetHelmRelease(this.endpointGuid, this.namespace, this.releaseTitle);
+    const entityService = this.esf.create<HelmRelease>(action.guid, action);
 
-    this.release$ = svc.entities$.pipe(
-      map((items: HelmRelease[]) => items.find(item => item.guid === this.guid)),
-      map((item: HelmRelease) => {
-        if (!item.chart.metadata.icon) {
-          const copy = JSON.parse(JSON.stringify(item));
-          copy.chart.metadata.icon = '/core/assets/custom/app_placeholder.svg';
-          return copy;
-        }
-        return item;
-      })
+
+    this.release$ = entityService.waitForEntity$.pipe(
+      map((item) => item.entity)
     );
+
+    // TODO: NWM This should use a `GetHelmRelease` action instead of fetching the entire list
+    // const paginationMonitor = new PaginationMonitor(store, action.paginationKey, kubernetesEntityFactory(helmReleaseEntityKey));
+    // const svc = getPaginationObservables({ store, action, paginationMonitor });
+    this.isFetching$ = entityService.isFetchingEntity$;
+
+    // this.release$ = svc.entities$.pipe(
+    //   map((items: HelmRelease[]) => items.find(item => item.guid === this.guid)),
+    //   map((item: HelmRelease) => {
+    //     if (!item.chart.metadata.icon) {
+    //       const copy = JSON.parse(JSON.stringify(item));
+    //       copy.chart.metadata.icon = '/core/assets/custom/app_placeholder.svg';
+    //       return copy;
+    //     }
+    //     return item;
+    //   })
+    // );
   }
 
   public guidAsUrlFragment(): string {
@@ -71,9 +78,9 @@ export class HelmReleaseHelperService {
 
   public fetchReleaseStatus(): Observable<HelmReleaseStatus> {
     // Get helm release
-    const action = new GetHelmReleaseStatus(this.endpointGuid, this.releaseTitle);
+    const action = new GetHelmRelease(this.endpointGuid, this.namespace, this.releaseTitle);
 
-    return this.esf.create<HelmReleaseStatus>(action.key, action).waitForEntity$.pipe(
+    return this.esf.create<HelmReleaseStatus>(action.guid, action).waitForEntity$.pipe(
       map(entity => entity.entity)
     );
   }
