+++ custom-src/frontend/app/custom/kubernetes/workloads/store/workloads.reducers.ts
@@ -1,10 +1,10 @@
-import { UPDATE_HELM_RELEASE_STATUS } from './workloads.actions';
+import { UPDATE_HELM_RELEASE } from './workloads.actions';
 
 const defaultState = {};
 
 export function helmReleaseReducer(state: any = defaultState, action) {
   switch (action.type) {
-    case UPDATE_HELM_RELEASE_STATUS:
+    case UPDATE_HELM_RELEASE:
       return {
         ...state,
       };
