+++ src/jetstream/plugins/kubernetes/main.go
@@ -119,6 +119,7 @@ func (c *KubernetesSpecification) AddSessionGroupRoutes(echoGroup *echo.Group) {
 	echoGroup.GET("/helm/releases", c.ListReleases)
 	echoGroup.POST("/helm/install", c.InstallRelease)
 	echoGroup.DELETE("/helm/releases/:endpoint/:namespace/:name", c.DeleteRelease)
+	echoGroup.GET("/helm/releases/:endpoint/:namespace/:name/status", c.GetReleaseStatus)
 	echoGroup.GET("/helm/releases/:endpoint/:namespace/:name", c.GetRelease)
 }
